pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kg-wiki"
version = "0.1.0"
description = "CLI framework for AI-maintained local knowledge graphs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["knowledge-graph", "second-brain", "llm", "rag", "notes"]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
kg = "kg.main:app"

[project.urls]
Repository = "https://github.com/ihsankhwaritsmi/knowledge-graph"

[tool.hatch.build.targets.wheel]
packages = ["src/kg"]

[tool.hatch.build.targets.wheel.shared-data]
"src/kg/templates" = "kg/templates"
